pyproject.toml
[tool.poetry]
name = "python-orm"
version = "0.1.0"
description = "Test ORM in Python"
authors = ["StephanMalan <stephanmalan.rob@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.6"
mysql-connector-python = "^8.0.33"


[tool.poetry.group.dev.dependencies]
coverage = "^7.2.5"
pytest = "^7.3.1"
black = "^23.3.0"
pylint = "^2.17.4"
mypy = "^1.3.0"
ruff = "^0.0.267"
types-psycopg2 = "^2.9.21.10"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.mypy]
exclude = ["test", "alembic"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
